# src/swagger.yaml
openapi: 3.0.1
info:
  title: Authentication API
  description: API for user authentication, including registration, login, logout, and token refresh functionality
  version: 1.0.0
  contact:
    name: Your Name
    email: your-email@example.com
    url: https://your-website.com
servers:
  - url: http://localhost:3000/auth
    description: Local server
paths:
...
  /refresh:
    post:
      summary: Refresh access and refresh tokens using the refresh token
      operationId: refresh
      requestBody:
        description: Refresh token to get new access and refresh tokens
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Successful token refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Invalid or expired refresh token
        '500':
          description: Server error
...